revert api changes, move queries to component, query client to api
