+++ src/App.tsx
@@ -2,7 +2,7 @@ import { createSignal, lazy, onCleanup, Show, Suspense, type ParentComponent, ty
 import { Router, Route } from '@solidjs/router'
 import { QueryClientProvider } from '@tanstack/solid-query'
 import { SolidQueryDevtools } from '@tanstack/solid-query-devtools'
-import { getAppQueryClient } from '~/queries/client'
+import { getAppQueryClient } from '~/api/query-client'
 
 import 'leaflet/dist/leaflet.css'
